MinerU/rag_test/local_rag_student/requirements.txt
ollama==0.1.6
python-docx==0.8.11
numpy==1.24.3
PyPDF2==3.0.1 